requirements.txt
streamlit>=1.35.0
python-jobspy>=1.1.80
anthropic>=0.25.0
python-docx>=1.1.0
pandas>=2.0.0
